move /internal/pathutil to /pkg/util/pathutil as per https://github.com/skycoin/skywire-updater/pull/7#pullrequestreview-227368978
