+++ cmd/skywire-node/commands/root.go
@@ -12,8 +12,8 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/skycoin/skywire/internal/pathutil"
 	"github.com/skycoin/skywire/pkg/node"
+	"github.com/skycoin/skywire/pkg/util/pathutil"
 )
 
 const configEnv = "SW_CONFIG"
