+++ pkg/manager/config.go
@@ -8,7 +8,7 @@ import (
 	"path/filepath"
 	"time"
 
-	"github.com/skycoin/skywire/internal/pathutil"
+	"github.com/skycoin/skywire/pkg/util/pathutil"
 
 	"github.com/skycoin/skywire/pkg/cipher"
 )
